+++ internal/goscrobble/timers.go
@@ -8,11 +8,10 @@ import (
 var endTicker chan bool
 
 func StartBackgroundWorkers() {
-	updateSpotifyData()
 
 	endTicker := make(chan bool)
 
-	hourTicker := time.NewTicker(time.Hour)
+	hourTicker := time.NewTicker(time.Duration(1) * time.Hour)
 	minuteTicker := time.NewTicker(time.Duration(60) * time.Second)
 
 	go func() {
@@ -25,6 +24,9 @@ func StartBackgroundWorkers() {
 				// Clear old password reset tokens
 				clearOldResetTokens()
 
+				// Attempt to pull missing images from spotify - hackerino version!
+				user, _ := getUserByUsername("idanoo")
+				user.updateImageDataFromSpotify()
 			case <-minuteTicker.C:
 				// Update playdata from spotify
 				updateSpotifyData()